requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
distlib==0.3.3
dnspython==2.1.0
filelock==3.0.12
Flask==2.0.1
flask-paginate==0.8.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
lxml==4.6.5
MarkupSafe==2.0.1
nltk==3.6.3
platformdirs==2.3.0
pymongo==3.12.0
pyspellchecker==0.6.2
python-dotenv==0.19.0
regex==2021.9.30
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
tqdm==4.62.3
urllib3==1.26.6
virtualenv==20.8.1
virtualenv-clone==0.5.7
Werkzeug==2.0.1